Add script to amd-ify the commonjs code, (backwards compat. and dynamic loading)
