+++ .gitignore
@@ -29,12 +29,13 @@ src
 .eggs/
 
 # Bundled/generated JavaScript
-notebook/static/auth/js/main.bundle.js
-notebook/static/edit/js/main.bundle.js
-notebook/static/notebook/js/main.bundle.js
-notebook/static/services/contents.bundle.js
-notebook/static/terminal/js/main.bundle.js
-notebook/static/tree/js/main.bundle.js
+notebook/static/auth/js
+notebook/static/base/js
+notebook/static/edit/js
+notebook/static/notebook/js
+notebook/static/services
+notebook/static/terminal/js
+notebook/static/tree/js
 
 # npm logs
 npm-debug.log
