+++ package.json
@@ -19,7 +19,7 @@
     "build:css:ipython": "lessc --source-map --include-path=./notebook/static/ ./notebook/static/style/ipython.less ./notebook/static/style/ipython.min.css",
     "build:css:style": "lessc --source-map --include-path=./notebook/static/ ./notebook/static/style/style.less ./notebook/static/style/style.min.css",
     "build:js": "concurrent \"npm run build:js:amd\" \"npm run build:js:tree\" \"npm run build:js:terminal\" \"npm run build:js:notebook\" \"npm run build:js:edit\" \"npm run build:js:auth\" \"npm run build:js:contents\"",
-    "build:amd": "node ./notebook/build.js",
+    "build:js:amd": "node ./notebook/amd.js",
     "build:js:notebook": "node ./notebook/build.js notebook/js/main.js notebook/js/main.bundle.js",
     "build:js:contents": "node ./notebook/build.js services/contents.js services/contents.bundle.js",
     "build:js:edit": "node ./notebook/build.js edit/js/main.js edit/js/main.bundle.js",
@@ -39,6 +39,7 @@
     "jquery-ui": "~1.10",
     "less": "~2",
     "marked": "~0.3",
+    "mkdirp": "^0.5.1",
     "moment": "~2.8.4",
     "rimraf": "^2.4.2",
     "term.js": "~0.0.4",
